+++ osk-disasm/rof.h
@@ -129,10 +129,10 @@ void RofLoadInitData(void);
 AddrSpaceHandle rof_class(int typ, int refTy);
 struct rof_extrn* find_extrn(refmap& xtrn, unsigned int adrs);
 int rof_datasize(char cclass, struct options* opt);
-void DataDoBlock(refmap* refsList, size_t blkEnd, AddrSpaceHandle space, struct parse_state* state);
+void DataDoBlock(refmap* refsList, uint32_t blkEnd, AddrSpaceHandle space, struct parse_state* state);
 void setupROFPass(int Pass);
 int rof_setup_ref(refmap& ref, int addrs, char* dest, int val);
-char* IsRef(char* dst, int curloc, int ival, int Pass);
+char* IsRef(char* dst, uint32_t curloc, int ival, int Pass);
 const char* extern_def_name(struct rof_extrn* handle);
 char* rof_header_getPsectParams(struct rof_header* handle);
